+++ packages/api-client/src/v2/blocks/scalar-auth-selector-block/index.ts
@@ -1,4 +1,5 @@
 export { default as AuthSelector } from './components/AuthSelector.vue'
 export { isAuthOptional } from './helpers/is-auth-optional'
 export { type MergedSecuritySchemes, mergeSecurity } from './helpers/merge-security'
+export type { OAuth2Tokens } from './helpers/oauth'
 export type { SecuritySchemeObjectSecret } from './helpers/secret-types'