+++ packages/api-reference/src/components/AgentScalar/index.ts
@@ -1,2 +1,3 @@
 export { default as AgentScalarButton } from './AgentScalarButton.vue'
 export { default as AgentScalarDrawer } from './AgentScalarDrawer.vue'
+export { default as OpenMCPButton } from './OpenMCPButton.vue'