+++ integrations/express/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @scalar/express-api-reference
 
+## 0.9.1
+
+### Patch Changes
+
+#### Updated Dependencies
+
+- **@scalar/core@0.4.1**
+
 ## 0.9.0
 
 ### Minor Changes