+++ packages/use-codemirror/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @scalar/use-codemirror
 
+## 0.14.3
+
+### Patch Changes
+
+#### Updated Dependencies
+
+- **@scalar/components@0.20.3**
+
+## 0.14.2
+
+### Patch Changes
+
+#### Updated Dependencies
+
+- **@scalar/components@0.20.2**
+  - [#8366](https://github.com/scalar/scalar/pull/8366): fix(components): clean up search components and add snapshots
+
+## 0.14.1
+
+### Patch Changes
+
+#### Updated Dependencies
+
+- **@scalar/components@0.20.1**
+
 ## 0.14.0
 
 ### Minor Changes